Add files via upload
This is a python code for our chainwheel robot with an ultrasonic sensor.
Superpanzerkampfwagen goes straight till it is closer to an object than 20cm. In this case Superpanzerkampfwagen turns to the right for 2 seconds, then continues to go forward again. Is it unstoppable? Well, yes, but technically no. If there's an object right in front of Superpanzerkampfwagen (less than 5cm) it stops like a well behaved dog.
Be careful, because Dr. End, the mad scientist will develop it humiliate everything.